Sorting and polspectra fix

+++ scripts/casda_prepare.py
@@ -2,12 +2,12 @@
 """Prepare files for CASDA upload"""
 import os
 import time
-from typing import Tuple, List
+from typing import Tuple, List, Dict
 from spiceracs.utils import try_mkdir, tqdm_dask, try_symlink
 import polspectra
 from glob import glob
 from astropy.io import fits
-from astropy.table import Table
+from astropy.table import Table, Column
 import astropy.units as u
 import numpy as np
 import logging as log
@@ -26,6 +26,7 @@ from astropy.visualization import (MinMaxInterval, SqrtStretch,
                                    ImageNormalize, ZScaleInterval, LogStretch)
 import astropy.units as u
 from radio_beam import Beam
+from tqdm.auto import tqdm
 
 def make_thumbnail(cube_f: str, cube_dir: str):
     cube = np.squeeze(fits.getdata(cube_f))
@@ -88,7 +89,7 @@ def convert_spectra(
     spectrum: str, 
     polcat: Table, 
     spec_dir: str = "."
-    ) -> Tuple[u.Quantity, u.Quantity, u.Quantity, np.ndarray]:
+    ) -> dict:
     """Convert a ascii spectrum to FITS
 
     Args:
@@ -167,20 +168,45 @@ def convert_spectra(
         "_FDFmodel.dat",
         "_RMSF.dat",
     )
+    fdf_clean_data = []
+    fdf_dirty_data = []
+    fdf_model_data = []
+    rmsf_data = []
+    phis_list = []
+    phis_long_list = []
+
     for suffix in suffixes:
         unit = u.Jy / u.beam / rmsf
 
+        is_rmsf = "RMSF" in suffix
+        is_clean = "clean" in suffix
+        is_dirty = "dirty" in suffix
+        is_model = "model" in suffix
         if suffix == "_RMSF.dat":
             unit = u.dimensionless_unscaled
+
 
         rm_file = spectrum.replace(".dat", suffix)
         full_rm_data = np.loadtxt(rm_file).T
         phis = full_rm_data[0] * u.rad / u.m ** 2
+        if is_rmsf:
+            phis_long_list.append(phis)
+        elif is_clean:
+            phis_list.append(phis)
         rm_q_data = full_rm_data[1]
         rm_u_data = full_rm_data[2]
         rm_data = np.array([rm_q_data, rm_u_data]) * unit
         # Add dummy axis for RA/DEC
         rm_data = rm_data[:, :, np.newaxis, np.newaxis]
+        if is_rmsf:
+            rmsf_data.append(rm_data)
+        elif is_clean:
+            fdf_clean_data.append(rm_data)
+        elif is_dirty:
+            fdf_dirty_data.append(rm_data)
+        elif is_model:
+            fdf_model_data.append(rm_data)
+
         # Create HDU
         hdu = fits.PrimaryHDU(data=rm_data.value)
         hdu.header["NAXIS"] = len(rm_data.shape)
@@ -216,7 +242,18 @@ def convert_spectra(
         log.info(f"Writing {rm_f} from {rm_file}")
         hdu.writeto(rm_f, overwrite=True)
 
-    return (freq, data, noise, np.array(gauss_id))
+    return dict(
+            freq=freq, 
+            data=data, 
+            noise=noise, 
+            phis=phis_list, 
+            phis_long=phis_long_list,
+            fdf_clean=fdf_clean_data,
+            fdf_dirty=fdf_dirty_data,
+            fdf_model=fdf_model_data,
+            rmsf=rmsf_data,
+            gauss_id=np.array(gauss_id)
+        )
 
 
 @delayed
@@ -261,10 +298,16 @@ def find_cubes(data_dir: str = ".") -> list:
 def make_polspec(
     casda_dir: str,
     polcat: Table,
-    freqs: np.ndarray,
-    data: np.ndarray,
-    noises: np.ndarray,
-    gauss_ids: np.ndarray,
+    freq_arr: np.ndarray,
+    data_arr: np.ndarray,
+    noise_arr: np.ndarray,
+    phis_arr: np.ndarray,
+    phis_long_arr: np.ndarray,
+    fdf_clean_arr: np.ndarray,
+    fdf_dirty_arr: np.ndarray,
+    fdf_model_arr: np.ndarray,
+    rmsf_arr: np.ndarray,
+    gauss_id_arr: np.ndarray,
 ) -> None:
     """Make a PolSpectra table
 
@@ -276,13 +319,20 @@ def make_polspec(
         noises (np.ndarray): Array of noise arrays
         gauss_ids (np.ndarray): Array of Gaussian IDs
     """
-    freq = freqs[0]
+    freq = freq_arr[0]
 
     # Sort everying by gauss_ids
-    sort_idx = np.argsort(gauss_ids)
-    data = data[sort_idx]
-    noises = noises[sort_idx]
-    gauss_ids = gauss_ids[sort_idx]
+    sort_idx = np.argsort(gauss_id_arr)
+    data = data_arr[sort_idx]
+    noises = noise_arr[sort_idx]
+    gauss_ids = gauss_id_arr[sort_idx]
+    phis = phis_arr[sort_idx]
+    phis_long = phis_long_arr[sort_idx]
+    fdf_clean = fdf_clean_arr[sort_idx]
+    fdf_dirty = fdf_dirty_arr[sort_idx]
+    fdf_model = fdf_model_arr[sort_idx]
+    rmsf = rmsf_arr[sort_idx]
+
     cat_idxs = []
     for gauss_id in gauss_ids:
         # Find the row
@@ -327,6 +377,126 @@ def make_polspec(
         leakage=polcat["leakage"],
         flux_type=polcat["flux_type"],
     )
+
+    rmsf_unit = u.def_unit("RMSF")
+    unit = u.Jy / u.beam / rmsf_unit
+    radms = u.radian / u.m**2
+
+    # Following same approach in polspectra
+    phi_column = Column(
+        name="faraday_depth",
+        unit=radms,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Faraday depth",
+    )
+    phi_column[:] = [x for x in phis]
+
+    phi_long_column = Column(
+        name="faraday_depth_long",
+        unit=radms,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Faraday depth",
+    )
+    phi_long_column[:] = [x for x in phis_long]
+
+    fdf_q_dirty_column = Column(
+        name="FDF_Q_dirty",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Dirty Stokes Q FDF",
+    )
+    fdf_q_dirty_column[:] = [x for x in fdf_dirty[:, 0]]
+
+    fdf_u_dirty_column = Column(
+        name="FDF_U_dirty",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Dirty Stokes U FDF",
+    )
+    fdf_u_dirty_column[:] = [x for x in fdf_dirty[:, 1]]
+
+    fdf_q_clean_column = Column(
+        name="FDF_Q_clean",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Clean Stokes Q FDF",
+    )
+    fdf_q_clean_column[:] = [x for x in fdf_clean[:, 0]]
+
+    fdf_u_clean_column = Column(
+        name="FDF_U_clean",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Clean Stokes U FDF",
+    )
+    fdf_u_clean_column[:] = [x for x in fdf_clean[:, 1]]
+
+    fdf_q_model_column = Column(
+        name="FDF_Q_model",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Model Stokes Q FDF",
+    )
+    fdf_q_model_column[:] = [x for x in fdf_model[:, 0]]
+
+    fdf_u_model_column = Column(
+        name="FDF_U_model",
+        unit=unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Model Stokes U FDF",
+    )
+    fdf_u_model_column[:] = [x for x in fdf_model[:, 1]]
+
+    rmsf_q_column = Column(
+        name="RMSF_Q",
+        unit=rmsf_unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Stokes Q RMSF",
+    )
+    rmsf_q_column[:] = [x for x in rmsf[:, 0]]
+
+    rmsf_u_column = Column(
+        name="RMSF_U",
+        unit=rmsf_unit,
+        dtype='object',
+        shape=(),
+        length=spectrum_table.Nrows,
+        description="Stokes U RMSF",
+    )
+    rmsf_u_column[:] = [x for x in rmsf[:, 1]]
+
+    spectrum_table.table.add_columns(
+        [
+            phi_column,
+            phi_long_column,
+            fdf_q_dirty_column,
+            fdf_u_dirty_column,
+            fdf_q_clean_column,
+            fdf_u_clean_column,
+            fdf_q_model_column,
+            fdf_u_model_column,
+            rmsf_q_column,
+            rmsf_u_column,
+        ]
+    )
     outf = os.path.join(casda_dir, "spice_racs_dr1_polspec.fits",)
     log.info(f"Writing {outf}")
     spectrum_table.write_FITS(outf, overwrite=True)
@@ -341,12 +511,16 @@ def convert_pdf(pdf_file: str, plots_dir:str, spec_dir: str) -> None:
     png_file = pdf_file.replace(".pdf", "")
     png_file = os.path.join(plots_dir, os.path.basename(png_file))
     log.info(f"Converting {pdf_file} to {png_file}")
+
     cmd = f"pdftoppm {pdf_file} {png_file} -png"
+    log.info(cmd)
     sp.run(cmd.split(), check=True)
     # Grr, pdftoppm doesn't preserve the file name
     actual_name = f"{png_file}-1.png"
     wanted_name = f"{png_file}.png"
-    os.rename(actual_name, wanted_name)
+    if os.path.exists(actual_name):
+        os.rename(actual_name, wanted_name)
+    assert os.path.exists(wanted_name)
 
     # Make thumbnail plot for CASDA - needs same name as FITS
     thumb = os.path.join(spec_dir, os.path.basename(wanted_name))
@@ -387,7 +561,7 @@ def find_plots(data_dir: str = ".") -> list:
     """
     cut_dir = os.path.join(data_dir, "cutouts")
     log.info(f"Globbing for plots in {cut_dir}")
-    plots = glob(os.path.join(os.path.join(cut_dir, "*"), "*.pdf"))
+    plots = glob(os.path.join(os.path.join(cut_dir, "RACS_*"), "*.pdf"))
     log.info(f"Found {len(plots)} plots")
     return plots
 
@@ -405,9 +579,11 @@ def main(
 
     log.info("Starting")
     log.info(f"Dask client: {client}")
-
     log.info(f"Reading {polcatf}")
     polcat = Table.read(polcatf)
+    df = polcat.to_pandas()
+    df = df.sort_values(["stokes_I_fit_flag", "snr_polint"], ascending=[True,False])
+    polcat = polcat[df.index.values]
     polcat.add_index("cat_id")
 
     casda_dir = os.path.join(data_dir, "casda")
@@ -452,7 +628,25 @@ def main(
                     for rid in rem_ids:
                         f.write(f"{rid}\n")
                 assert len(cubes) == len(set(polcat["source_id"])) * 3 * 2
-        for cube in cubes:
+
+        unique_ids, unique_idx = np.unique(polcat["source_id"], return_index=True)
+        lookup = {sid: i for sid, i in zip(unique_ids, unique_idx)}
+        with tqdm(total=len(cubes), desc="Sorting cubes") as pbar:
+            def my_sorter(x, lookup=lookup, pbar=pbar):
+                basename = x.split("/")[-1]
+                cut_idx = basename.find(".cutout")
+                sourcename = basename[:cut_idx]
+                idx = lookup[sourcename]
+                if "weight" in basename:
+                    idx += 3
+                for i, s in enumerate((".i.", ".q.", ".u.")):
+                    if s in basename:
+                        idx += i
+                        pbar.update(1)
+                        return idx
+            sorted_cubes = sorted(cubes, key=my_sorter)
+
+        for cube in sorted_cubes:
             out = update_cube(cube=cube, cube_dir=cube_dir)
             cube_outputs.append(out)
 
@@ -463,8 +657,20 @@ def main(
         try_mkdir(spec_dir)
         spectra = find_spectra(data_dir=data_dir)
         assert len(spectra) == len(polcat) # Sanity check
+
+        unique_ids, unique_idx = np.unique(polcat["cat_id"], return_index=True)
+        lookup = {sid: i for sid, i in zip(unique_ids, unique_idx)}
+        with tqdm(total=len(spectra), desc="Sorting spectra") as pbar:
+            def my_sorter(x, lookup=lookup, pbar=pbar):
+                basename = x.split("/")[-1]
+                sourcename = basename.replace(".dat", "")
+                idx = lookup[sourcename]
+                pbar.update(1)
+                return idx
+            sorted_spectra = sorted(spectra, key=my_sorter)
+
         # Loop over spectra and convert to FITS
-        for spectrum in spectra:
+        for spectrum in sorted_spectra:
             # freq, datum, noise, gauss_id = convert_spectra(spectrum, spec_dir=spec_dir)
             out = convert_spectra(
                 spectrum=spectrum, 
@@ -481,7 +687,21 @@ def main(
         spec_dir = os.path.join(casda_dir, "spectra")
         try_mkdir(spec_dir)
         plots = find_plots(data_dir=data_dir)
-        for plot in plots:
+        unique_ids, unique_idx = np.unique(polcat["cat_id"], return_index=True)
+        lookup = {sid: i for sid, i in zip(unique_ids, unique_idx)}
+        with tqdm(total=len(plots), desc="Sorting plots") as pbar:
+            def my_sorter(x, lookup=lookup, pbar=pbar):
+                fname = x.split("/")[-1]
+                for plot_type in ("_spectra", "_RMSF", "_clean"):
+                    if plot_type in fname:
+                        sourcename = fname[:fname.find(plot_type)]
+                        break
+                idx = lookup[sourcename]
+                pbar.update(1)
+                return idx
+
+            sorted_plots = sorted(plots, key=my_sorter)
+        for plot in sorted_plots:
             out = convert_pdf(pdf_file=plot, plots_dir=plots_dir, spec_dir=spec_dir)
             plot_outputs.append(out)
 
@@ -497,6 +717,7 @@ def main(
             log.info(f"Testing {len(outputs)} {name}")
         else:
             log.info(f"Starting work on {len(outputs)} {name}")
+
         futures = client.persist(outputs)
         # dumb solution for https://github.com/dask/distributed/issues/4831
         log.debug("I sleep!")
@@ -506,26 +727,24 @@ def main(
         # For spectra, we also want to make a polspec catalogue
         if name == "spectra" and len(outputs) > 0:
             log.info("Making polspec catalogue")
-            freqs, data, noises, gauss_ids = [], [], [], []
+            keys = futures[0].compute().keys()
+            out_data_lists = {
+                key: [] for key in keys
+            } # type: Dict[str, list]
+
             for future in futures:
                 result = future.compute()
-                freqs.append(result[0])
-                data.append(result[1])
-                noises.append(result[2])
-                gauss_ids.append(result[3])
-            freqs_arr = np.squeeze(np.array(freqs))
-            data_arr = np.squeeze(np.array(data))
-            noises_arr = np.squeeze(np.array(noises))
-            gauss_ids_arr = np.squeeze(np.array(gauss_ids))
-
+                for key in keys:
+                    out_data_lists[key].append(result[key])
+            
+            out_data_arrs = {} # type: Dict[str, np.ndarray]
+            for key in keys:
+                out_data_arrs[f"{key}_arr"] = np.squeeze(np.array(out_data_lists[key]))
             # Make polspec catalogue
             make_polspec(
                 casda_dir=casda_dir,
                 polcat=polcat,
-                freqs=freqs_arr,
-                data=data_arr,
-                noises=noises_arr,
-                gauss_ids=gauss_ids_arr,
+                **out_data_arrs,
             )
 
     log.info("Done")
